requirements.txt
flask
flask-cors
python-dotenv
mysql-connector-python
scikit-learn
xgboost
pandas
numpy
